Commit for push

+++ src/main/java/kr/co/mz/order_paper/repository/OrderPaperRepository.java
@@ -3,6 +3,6 @@ package kr.co.mz.order_paper.repository;
 import kr.co.mz.order_paper.entity.OrderPaperEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderPaperRepository extends JpaRepository<OrderPaperEntity, Integer> {
+public interface OrderPaperRepository extends JpaRepository<OrderPaperEntity, Long> {
 
 }
